requirements.txt
streamlit
langchain
langchain-google-genai
python-dotenv
youtube-transcript-api